Removed Default Allocate Mechanism for Tasks
The first task was checked for having a resource allocation. If none
was found, it would default to the first resource, if no resource was
defined, it would defaul to the user name.

This has been removed as it does not export the org file but adds non
configurable defaults to it.